bb: add apply-labels.js module
This module examines the bucketized issues and applies the appropriate
labels. For each issue, this may result in no action (already has the
right label) or some combination of adding and removing labels.

This also adds the github module to actually do the labelling. We now
need a config and an example will be provided in the next commit.