Backlog balance: deal with issues in other repos
We now have issues from a bunch of other repos in pulldasher / our DB.
This query needed to be limited to only care about the ones on our
main repo. Otherwise it would consider all issues across all repos
(that have p-labels) as part of the backlog and the result would
be individual per-repo backlogs wouldn't be balanced.