Previous Priority: parse label events
Parse the issue events to retrieve the "previous" priority because
we decided this should be included in our sorting algorithm.

Given that priority labels are not mutually exclusive, it's not
trivial to give a result that is sensible 100% of the time.

I've opted for something that's correct *almost* always, and it much
easier to understand.

Current priority = the current priority label
Previous priority = the most recently removed priority label

The only time this might give odd results if current = p3 and for some
reason, someone accidentally briefly adds and then removes another
priority label.

time ->
+p3 ------------------------>
         +p2-->

A human would say that short-lived p2 was a mistake and should be
ignored, but this code will pick it up as the "previous".

This is a rare case and even if it is hit, it won't matter that much.